requirements.txt
# MEDEA-NEUMOUSA: Classical Studies AI Platform
# Core dependencies for the Talos Destroyer

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
jinja2==3.1.2
python-multipart==0.0.6

# AI/ML Core - The Nine Muses' Power
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
google-generativeai>=0.3.2

# Classical Studies - Medea's Ancient Knowledge
cltk>=1.1.6
greek-accentuation>=1.2.0
pypi-cltk>=1.0.0

# Modern NLP - Contemporary Sorcery
spacy>=3.7.0
nltk>=3.8.1

# Graph Analysis - Network of the Dead
networkx>=3.2.1
python-igraph>=0.11.0

# Vector Databases - Repository of Souls
chromadb>=0.4.15
pinecone-client>=2.2.4

# Caching & Storage - Memory Palace
redis>=5.0.1
aioredis>=2.0.1

# Data Processing - Textual Alchemy
pandas>=2.1.0
polars>=0.19.0

# Visualization - Seeing the Unseen
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities - Tools of the Trade
aiofiles>=23.2.1
python-multipart>=0.0.6
httpx>=0.25.0
asyncio-throttle>=1.0.2

# Development - Forge of Creation
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
isort>=5.12.0
mypy>=1.6.0

# Production - Battle Ready
gunicorn>=21.2.0
prometheus-client>=0.18.0

# Additional Classical Tools
pyuca>=1.2  # Unicode Collation Algorithm for ancient scripts
regex>=2023.10.3  # Advanced regex for complex ancient text patterns